Unify the test frameworks for running loops
